update to dynamic route
